+++ inst/examples/spinner2.R
@@ -1,0 +1,28 @@
+#'
+# /*
+if (interactive()) {
+  devtools::load_all()
+} else {
+# */
+library(qwraps2)
+packageVersion("qwraps2")
+# /*
+}
+# */
+#'
+#' # mtcars2
+# /* {{{ */
+#'
+#' some blah blah blah
+#'
+# /* }}} */
+#'
+#'
+#+ label = "chunk1"
+2 + 2
+
+#'
+#'
+#' blah blah
+#'
+#'